% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{testCoverage-package}
\alias{testCoverage-package}
\title{testCoverage Package}
\description{
testCoverage is a package to report on the code coverage provided by a suite
of unit tests written in either testthat or RUnit.
}
\details{
HTML report output annotates source code with the coverage provided by
various unit test files.
}
\examples{
\dontrun{
demo(saturation)}
}
\author{
Mango Solutions\email{support@mango-solutions.com}
}
\seealso{
\code{\link{testCoverage}}, \code{\link{pkgCoverage}}, \code{\link{cranCoverage}}, \code{\link{selfCoverage}} and \code{vignette("testCoverage", package = "testCoverage")}
}

